design: 스켈레톤UI, NotFound 페이지 생성
