+++ public/index.html
@@ -18,7 +18,7 @@
     <!-- 카카오맵 라이브러리 -->
     <script
       type="text/javascript"
-      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=188be1d87e6004afccadc3cb7c01351a&libraries=services,clusterer"
+      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=1e25c92264f75a16d289d2a8216ddbbe&libraries=services,clusterer"
     ></script>
     <!-- 리셋 css -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/reset-css@5.0.1/reset.min.css" />
