+++ src/components/shared/Loading/Loading/Loading.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './Loading.scss';
+function Loading() {
+  return (
+    <div className="loading-box">
+      <svg class="spinner" width="65px" height="65px" viewBox="0 0 66 66" xmlns="http://www.w3.org/2000/svg">
+        <circle class="path" fill="none" stroke-width="6" stroke-linecap="round" cx="33" cy="33" r="30"></circle>
+      </svg>
+    </div>
+  );
+}
+
+export default Loading;
