+++ src/shared/Router.js
@@ -4,7 +4,11 @@ import Header from '../components/shared/Header';
 import Search from '../pages/Search';
 import StoreDetail from '../pages/StoreDetail';
 import MyPage from '../pages/MyPage';
-import NotFound from '../components/shared/NotFound';
+
+// NotFound 추가됨
+import NotFound from '../components/shared/NotFound/NotFound';
+import Loading from '../components/shared/Loading/Loading/Loading';
+import SkeletonUi from '../components/shared/Loading/SkeletonUi/SkeletonUi';
 
 const Router = () => {
   return (
@@ -16,6 +20,8 @@ const Router = () => {
         <Route path="/store/:id" element={<StoreDetail />} />
         <Route path="/mypage" element={<MyPage />} />
         <Route path="/*" element={<NotFound />} />
+        <Route path="/aa1" element={<Loading />} />
+        <Route path="/aa2" element={<SkeletonUi />} />
       </Routes>
     </BrowserRouter>
   );
